@@ -3,8 +3,9 @@
   "version": "0.9.66",
   "private": true,
   "dependencies": {
+    "@chakra-ui/css-reset": "^1.1.3",
     "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.6",
+    "@chakra-ui/react": "^1.8.7",
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
     "@fortawesome/fontawesome": "^1.1.8",
@@ -14,16 +15,16 @@
     "@fortawesome/react-fontawesome": "^0.1.18",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.23",
-    "@types/react": "^17.0.42",
+    "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
     "countries-and-timezones": "^3.3.0",
     "dayjs": "^1.11.0",
-    "framer-motion": "^4",
+    "framer-motion": "^6",
     "friendly-mimes": "^1.0.5",
     "javascript-time-ago": "^2.3.13",
     "path-browserify": "^1.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-dropzone": "^12.0.4",
     "react-markdown": "^8.0.1",
     "react-mde": "^11.5.0",
@@ -32,14 +33,14 @@
     "react-select": "^5.2.2",
     "react-time-ago": "^7.1.9",
     "recharts": "^2.1.9",
-    "typescript": "^4.6.2",
+    "typescript": "^4.6.3",
     "ua-parser-js": "^1.0.2",
     "xterm": "^4.18.0"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/user-event": "^14.0.0",
     "npm-check-updates": "^12.5.4",
     "sass": "^1.49.9"
   },
